# apps/inference/openapi.json
{
  "components": {
    "schemas": {
...
      "ActivationSingleBatchRequest": {
        "description": "Get activations for either a specific feature in an SAE (specified by \"source\" + \"index\") or a custom vector (specified by \"vector\" + \"hook\")",
        "properties": {
          "hook": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Hook that the custom vector applies to. Must be specified with \"vector\".",
            "title": "Hook"
          },
          "index": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Index of the SAE. Must be specified with \"source\", or not at all.",
            "title": "Index"
          },
          "model": {
            "description": "Name of the model to test activations on",
            "title": "Model",
            "type": "string"
          },
          "prompts": {
            "description": "Input text prompts to get activations for",
            "items": {
              "type": "string"
            },
            "title": "Prompts",
            "type": "array"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Source identifier - could be an SAE ID (eg 5-gemmascope-res-16k). Must be specified with \"index\", or not at all.",
            "title": "Source"
          },
          "vector": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Custom vector to test activations. Must be specified with \"hook\".",
            "title": "Vector"
          }
        },
        "required": [
          "prompts",
          "model"
        ],
        "title": "ActivationSingleBatchRequest",
        "type": "object"
      },
      "ActivationSingleBatchResponse": {
        "description": "Response for NPActivationBatchRequest. Contains the batch results of activation values and tokenized prompt.",
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ActivationSingleBatchResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "ActivationSingleBatchResponse",
        "type": "object"
      },
      "ActivationSingleBatchResult": {
        "description": "One prompt's result within a batched single-feature request.",
        "properties": {
          "activation": {
            "$ref": "#/components/schemas/ActivationValues"
          },
          "tokens": {
            "items": {
              "type": "string"
            },
            "title": "Tokens",
            "type": "array"
          },
          "tokensIsSpecial": {
            "description": "Index-aligned with \"tokens\". True where the token is one of the tokenizer's special tokens (BOS, EOS, padding, chat turn markers, and so on) rather than content. Lets a client filter scaffolding without knowing any model family's token literals.",
            "items": {
              "type": "boolean"
            },
            "title": "Tokensisspecial",
            "type": "array"
          }
        },
        "required": [
          "activation",
          "tokens",
          "tokensIsSpecial"
        ],
        "title": "ActivationSingleBatchResult",
        "type": "object"
      },
...
      "ActivationTopkByTokenBatchRequest": {
        "description": "Get activations for either a specific feature in an SAE (specified by \"source\" + \"index\") or a custom vector (specified by \"vector\" + \"hook\")",
        "properties": {
          "ignoreBos": {
            "description": "Whether or not to include features whose highest activation value is the BOS token.",
            "title": "Ignorebos",
            "type": "boolean"
          },
          "model": {
            "description": "Name of the model to test activations on",
            "title": "Model",
            "type": "string"
          },
          "prompts": {
            "description": "Input text prompts to get activations for",
            "items": {
              "type": "string"
            },
            "title": "Prompts",
            "type": "array"
          },
          "source": {
            "description": "Source identifier - could be an SAE ID (eg 5-gemmascope-res-16k). Must be specified with \"index\", or not at NPActivationAllRequest.",
            "title": "Source",
            "type": "string"
          },
          "topK": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "The number of features to include for each token position.",
            "title": "Topk"
          }
        },
        "required": [
          "prompts",
          "model",
          "source",
          "ignoreBos"
        ],
        "title": "ActivationTopkByTokenBatchRequest",
        "type": "object"
      },
      "ActivationTopkByTokenBatchResponse": {
        "description": "Response for NPActivationTopkByTokenBatchRequest. Contains the batch results of top features at each token position and the tokenized prompts.",
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ActivationTopkByTokenBatchResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "ActivationTopkByTokenBatchResponse",
        "type": "object"
      },
      "ActivationTopkByTokenBatchResult": {
        "description": "One prompt's result within a batched topk-by-token request.",
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ActivationTopkByTokenResult"
            },
            "title": "Results",
            "type": "array"
          },
          "tokens": {
            "items": {
              "type": "string"
            },
            "title": "Tokens",
            "type": "array"
          }
        },
        "required": [
          "results",
          "tokens"
        ],
        "title": "ActivationTopkByTokenBatchResult",
        "type": "object"
      },
      "ActivationTopkByTokenFeature": {
        "description": "One feature active at a single token position.",
        "properties": {
          "activationValue": {
            "description": "The activation value of this feature at this token position.",
            "title": "Activationvalue",
            "type": "number"
          },
          "featureIndex": {
            "description": "The index of the feature in the SAE.",
            "title": "Featureindex",
            "type": "integer"
          }
        },
        "required": [
          "featureIndex",
          "activationValue"
        ],
        "title": "ActivationTopkByTokenFeature",
        "type": "object"
      },
...
      "ActivationTopkByTokenResult": {
        "description": "One token's TopK result, including its top features.",
        "properties": {
          "isSpecial": {
            "description": "Whether this token is one of the tokenizer's special tokens (BOS, EOS, padding, chat turn markers, and so on) rather than content. Lets a client filter scaffolding without knowing any model family's token literals.",
            "title": "Isspecial",
            "type": "boolean"
          },
          "token": {
            "description": "The token string",
            "title": "Token",
            "type": "string"
          },
          "tokenPosition": {
            "description": "The index of the token in the prompt.",
            "title": "Tokenposition",
            "type": "integer"
          },
          "topFeatures": {
            "items": {
              "$ref": "#/components/schemas/ActivationTopkByTokenFeature"
            },
            "title": "Topfeatures",
            "type": "array"
          }
        },
        "required": [
          "tokenPosition",
          "token",
          "isSpecial",
          "topFeatures"
        ],
        "title": "ActivationTopkByTokenResult",
        "type": "object"
      },
      "ActivationValues": {
        "description": "One feature's activation over a prompt, plus its DFA attribution if computed.",
        "properties": {
          "dfaMaxValue": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dfamaxvalue"
          },
          "dfaTargetIndex": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dfatargetindex"
          },
          "dfaValues": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dfavalues"
          },
          "maxValue": {
            "title": "Maxvalue",
            "type": "number"
          },
          "maxValueIndex": {
            "title": "Maxvalueindex",
            "type": "integer"
          },
          "values": {
            "items": {
              "type": "number"
            },
            "title": "Values",
            "type": "array"
          }
        },
        "required": [
          "values",
          "maxValue",
          "maxValueIndex"
        ],
        "title": "ActivationValues",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "SimpleSecretAuth": {
        "in": "header",
        "name": "X-SECRET-KEY",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "email": "johnny@neuronpedia.org"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Neuronpedia - Inference Server",
    "version": "1.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/activation/single-batch": {
      "post": {
        "operationId": "activationSingleBatchPost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivationSingleBatchRequest",
                "examples": [
                  {
                    "index": "14057",
                    "model": "gpt2-small",
                    "prompts": [
                      "The Jedi in Star Wars wield lightsabers.",
                      "The Force is strong with this one."
                    ],
                    "source": "0-res-jb"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivationSingleBatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activation Single Batch"
      }
    },
...
    "/v1/activation/topk-by-token-batch": {
      "post": {
        "operationId": "activationTopkByTokenBatchPost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivationTopkByTokenBatchRequest",
                "examples": [
                  {
                    "ignore_bos": true,
                    "model": "gpt2-small",
                    "prompts": [
                      "The Jedi in Star Wars wield lightsabers.",
                      "The Force is strong with this one."
                    ],
                    "source": "0-res-jb"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivationTopkByTokenBatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activation Topk By Token Batch"
      }
    },
...
  },
  "security": [
    {
      "SimpleSecretAuth": []
    }
  ]
}